% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_clean_raw}
\alias{sap_clean_raw}
\title{Convenience method to top and tail (remove firat and last column and row from raw SAP output format}
\usage{
sap_clean_raw(o)
}
\arguments{
\item{o}{object to clean}
}
\description{
Convenience method to top and tail (remove firat and last column and row from raw SAP output format
}
\examples{
i <- sap_clean_raw(o)
}
\keyword{SAP}
\keyword{unformated}
