% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_date}
\alias{sap_date}
\title{A Function To Clean and Convert SAP Dates
The function replace , and converts to dates}
\usage{
sap_date(f)
}
\arguments{
\item{f}{The value to be converted}

\item{l}{The lubridate pattern, optional}
}
\description{
A Function To Clean and Convert SAP Dates
The function replace , and converts to dates
}
\examples{
i <- sap_date(date_value,"ymd")
}
\keyword{SAP}
\keyword{conversion}
\keyword{date}
