% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_duration_calc}
\alias{sap_duration_calc}
\title{Calculate Duration Based On Separate Date and Time Fields For Start And End
The function calculates duration}
\usage{
sap_duration_calc(start_date, end_date, start_time, end_time)
}
\arguments{
\item{start_date}{start date}

\item{end_date}{end date}

\item{start_time}{start time}

\item{end_time}{end time}
}
\description{
Calculate Duration Based On Separate Date and Time Fields For Start And End
The function calculates duration
}
\examples{
i <- sap_duration_calc(st_dt, en_dt, st_tm, en_tm)
}
\keyword{Duration}
\keyword{SAP}
