% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_file_load}
\alias{sap_file_load}
\title{A Function To Read STAD Unformated Exports
The function reads SAP STAD exports}
\usage{
sap_file_load(dir_name, pattern, cleanup, skip)
}
\arguments{
\item{dir_name}{The directory containing the STAD exports}

\item{pattern}{The file pattern to read}

\item{cleanup}{A function to clean read data}

\item{skip}{The nuumber of lines to skip}
}
\description{
A Function To Read STAD Unformated Exports
The function reads SAP STAD exports
}
\examples{
i <- sap_file_load("directory",".*",my_cleanup,4)
}
\keyword{SAP}
\keyword{STAD}
