% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_find_skip}
\alias{sap_find_skip}
\title{A Function To Determine Lines To Skip Base On First Line Starting With A Character
The function returns the first line that starts with a particular separator
The default is ten lines to check and for the pipe (|) character}
\usage{
sap_find_skip(f, l = 10, c = "|")
}
\arguments{
\item{f}{filename}

\item{l}{lines}

\item{c}{character}
}
\description{
A Function To Determine Lines To Skip Base On First Line Starting With A Character
The function returns the first line that starts with a particular separator
The default is ten lines to check and for the pipe (|) character
}
\examples{
sap_find_skip(filename)
}
\keyword{skip}
