% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_missing_values}
\alias{sap_missing_values}
\title{A Function To Form A Data Frame Of Missing Values
The function gives a data frame of columns that have missing data and the quantity in each}
\usage{
sap_missing_values(o)
}
\arguments{
\item{o}{The object to be evaluated}
}
\description{
A Function To Form A Data Frame Of Missing Values
The function gives a data frame of columns that have missing data and the quantity in each
}
\examples{
sap_stad_quantile(stad_values, field, filter_value, summaryfield)
}
\keyword{na}
