% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_numeric}
\alias{sap_numeric}
\title{A Function To Clean and Convert SAP Numbers
The function replace , and converts to numeric}
\usage{
sap_numeric(f)
}
\arguments{
\item{f}{The value to be converted}
}
\description{
A Function To Clean and Convert SAP Numbers
The function replace , and converts to numeric
}
\examples{
i <- sap_numeric(stad_value)
}
\keyword{SAP}
\keyword{conversion}
\keyword{number}
