% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_stad_quantile}
\alias{sap_stad_quantile}
\title{A Function To Calculate A Vector Of Quantile Values
The function calculates quantile values such as from STAD data}
\usage{
sap_stad_quantile(o, testfield, pattern, summaryfield)
}
\arguments{
\item{o}{The object to be evaluated}

\item{testfield}{The field to be calculated for}

\item{pattern}{A filter pattern for the textfield}

\item{summaryfield}{The field to be summarised}
}
\description{
A Function To Calculate A Vector Of Quantile Values
The function calculates quantile values such as from STAD data
}
\examples{
sap_stad_quantile(stad_values, field, filter_value, summaryfield)
}
\keyword{SAP}
\keyword{STAD}
\keyword{quantile}
