% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_tidy_sdf}
\alias{sap_tidy_sdf}
\title{A Starter Function To Clean SDF MON header data
The function clean SAP SDF MON header exports}
\usage{
sap_tidy_sdf(o)
}
\arguments{
\item{o}{The dataframe to clean}
}
\description{
A Starter Function To Clean SDF MON header data
The function clean SAP SDF MON header exports
}
\examples{
i <- sap_tidy_sdf(df)
}
\keyword{MON}
\keyword{SAP}
\keyword{SDF}
