% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_tidy_st03n_hr}
\alias{sap_tidy_st03n_hr}
\title{A Starter Function To Clean ST03N Hourly Data
The function clean SAP ST03N hourly export exports}
\usage{
sap_tidy_st03n_hr(o)
}
\arguments{
\item{o}{The dataframe to clean}
}
\description{
A Starter Function To Clean ST03N Hourly Data
The function clean SAP ST03N hourly export exports
}
\examples{
i <- sap_tidy_st03n_hr(df)
}
\keyword{SAP}
\keyword{ST03N}
