% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_tidy_stad}
\alias{sap_tidy_stad}
\title{A Starter Function To Clean STAD data
The function clean SAP STAD exports first field}
\usage{
sap_tidy_stad(f)
}
\arguments{
\item{f}{The dataframe to clean}
}
\description{
A Starter Function To Clean STAD data
The function clean SAP STAD exports first field
}
\examples{
i <- sap_tidy_stad(df)
}
\keyword{SAP}
\keyword{STAD}
