% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_general_functions.R
\name{sap_time}
\alias{sap_time}
\title{A Function To Clean and Convert SAP Times in the format HHMMSS}
\usage{
sap_time(f)
}
\arguments{
\item{f}{The value to be converted}
}
\description{
A Function To Clean and Convert SAP Times in the format HHMMSS
}
\examples{
i <- sap_time("083000")
}
\keyword{SAP}
\keyword{conversion}
\keyword{time}
