% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_timeline.R
\name{sap_timeline}
\alias{sap_timeline}
\title{A Timeline function}
\usage{
sap_timeline(o, testfield, timefield)
}
\arguments{
\item{o}{The object to be evaluated}

\item{testfield}{The field to be monitored for changes within the samples}

\item{timefield}{The sample datetime field}
}
\description{
The function calculates logical start and end times for an observation to support timeline drawing.
The timeline itself is draw by subsequent processing of the returned data.
}
\examples{
timeline_data  <- sap_timeline(rawdata,'OBJECT_NAME','dt')

timeline_data \%>\% group_by(start_time,OBJECT_NAME) \%>\% ggplot() + geom <- segment(aes(x=start <- time, xend=dt,y=OBJECT_NAME, yend=OBJECT_NAME,colour=OBJECT_NAME),size=6)+theme(legend.position='none')+labs(title='Sample Timeline')
}
\keyword{SAP}
\keyword{timeline}
